refactor(manifest): move the test module to a sibling file
`manifest.rs` read as a 5,293-line file. 3,761 of those were an inline
`mod tests`; the parser itself is 1,533.

A pure move, plus the eight `include_str!` paths that had to gain a `../`
because they resolve relative to the file rather than the crate. The sibling
layout is also what CONTRIBUTING names as the sanctioned shape for a test
module large enough to obscure the code it tests.

761 tests, unchanged, still passing.
